celery_worker/requirements.txt
celery==5.3.6
flower==1.0.0
redis==3.5.3
watchdog==2.1.5
python-dotenv==0.19.0
charset-normalizer==3.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
rsa==4.9
tqdm==4.66.1
urllib3==2.1.0
